Initial setup of app with typescript and redux.
This commit has the initial setup of a Create React App typescript and redux starter. Docker is also setup to build the app and run it on an nginx server. The database file is added and server setup using `json-server` and an environment variable has been created for the api url.

Task 1.1  Task1.2